move prisma client
