+++ back-end/src/client.ts
@@ -1,0 +1,7 @@
+import { PrismaClient } from './.generated/prisma';
+
+const prisma = new PrismaClient({
+  log: ['info', 'warn', 'error'],
+});
+
+export default prisma;
