+++ back-end/src/story.ts
@@ -1,6 +1,7 @@
 import { Router } from 'express';
 
 import { Category, Game, GamePhase, StoryEntry } from './.generated/prisma';
+import prisma from './client';
 import {
   StoryArchiveResBody as ArchiveRes,
   StoryReqBody,
@@ -8,7 +9,6 @@ import {
   UserDto,
 } from './domain/types.js';
 import { joinPhase, loadStory, loadUser } from './middleware.js';
-import prisma from './server.js';
 import { WAIT, punctRegex, quoteRegex } from './utils/constants.js';
 import { ReqHandler as Handler, ReqBody as ReqBody } from './utils/types.js';
 import {
